audio/requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
Django==5.0.1
django-debug-toolbar==4.2.0
django-taggit==5.0.1
djangorestframework==3.13.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pytz==2023.4
sqlparse==0.4.4
